pyproject.toml
[project]
name = "neural-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "sam3690",email = "usamabinayoub@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "groq (>=0.26.0,<0.27.0)",
    "ipython (>=9.2.0,<10.0.0)"
]

[tool.poetry]
packages = [{include = "neural_agent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
